Complete Stage 3/8: Palindromic numbers
Add a method to determine if a number is a palindrome or not (reads the same left to right as it does right to left).

Implement a loop so we now keep asking for numbers until the user enters 0 to exit.

Change error messages to match requirements of the stage.